requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
python-dotenv
openai
pyupbit
ta
selenium
youtube-transcript-api
pandas_market_calendars
firebase-admin==6.4.0